+++ m/easers.lua
@@ -17,7 +17,7 @@ function m.quad_out(k)
   return k * (2 - k)
 end
 
-function m.in_out(k)
+function m.quad_in_out(k)
   k = k * 2
   if k < 1 then
     return 0.5 * k * k;
@@ -81,7 +81,7 @@ function m.quint_in_out(k)
 end
 
 function m.sine_in(k)
-  return 1 - math.sin(k * math.pi / 2)
+  return 1 - math.cos(k * math.pi / 2)
 end
 
 function m.sine_out(k)
@@ -89,7 +89,7 @@ function m.sine_out(k)
 end
 
 function m.sine_in_out(k)
-  return 0.5 * (1 - math.cos(math.pi * 2))
+  return 0.5 * (1 - math.cos(math.pi * k))
 end
 
 function m.expo_in(k)